requirements.txt
# Core dependencies
flask==3.0.0
flask-cors==4.0.0
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1

# Spotify and Last.fm APIs
spotipy==2.23.0
pylast==5.2.0
requests==2.31.0
python-dotenv==1.0.0

# Data processing
pandas==2.1.3
numpy==1.26.2
pytz==2023.3

# Scheduling
apscheduler==3.10.4

# Visualization
plotly==5.18.0
dash==2.14.1
matplotlib==3.8.2
seaborn==0.13.0
calplot==0.1.7.4

# Configuration
pyyaml==6.0.1

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0

# Logging
loguru==0.7.2